remove `v` or `t` from identifer list
There are many files named `{funcname}.f.html` or `{traitname}.t.html`
in document directory.  However these are displayed as `f` or `v`
(according to current matching regexp) so there are many `f`s and `v`s
in the identifier list.  This is meaningless.  Moreover, these file is
simply a link to `fn.{funcname}.html`. So these name should be filtered
out.